New version with static models
I estimated all state space models on data ranging form 2020-09-15 to 2021-04-25 and saved tema as a Rdata file.
When the app starts it loads data and models and run the smoother using new data and previously estimated models.
Last-day data can now be seen also in a scatterplot, however I would like to hide the tab of the scatterplot when Nationale is selected. This is a fix to do.